% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_gibbs}
\alias{arms_gibbs}
\title{Multivariate ARMS via Gibbs sampler}
\usage{
arms_gibbs(
  n,
  initialPoints,
  minRange,
  maxRange,
  metropolis,
  betaJ,
  vA,
  vB,
  datX0,
  datProportion,
  datEvent,
  weibullS
)
}
\arguments{
\item{n}{number of variates to draw}

\item{initialPoints}{this can be a matrix in multivariate case}
}
\description{
Multivariate ARMS via Gibbs sampler
}
