% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_gibbs_zeta}
\alias{arms_gibbs_zeta}
\title{Multivariate ARMS via Gibbs sampler for zeta}
\usage{
arms_gibbs_zeta(
  n,
  nsamp,
  ninit,
  minRange,
  maxRange,
  metropolis,
  simple,
  convex,
  npoint,
  currentPars,
  w0Sq,
  wSq,
  phi,
  kappa,
  dirichlet,
  datX,
  datTheta,
  datProportionConst,
  datEvent,
  weibullS,
  weibullLambda
)
}
\arguments{
\item{n}{Number of samples to draw}

\item{nsamp}{How many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}

\item{dirichlet}{Not yet implemented}
}
\description{
Multivariate ARMS via Gibbs sampler for zeta
}
