% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_kappa}
\alias{arms_kappa}
\title{Univariate ARMS for kappa}
\usage{
arms_kappa(
  n,
  nsamp,
  ninit,
  minRange,
  maxRange,
  metropolis,
  simple,
  convex,
  npoint,
  currentPars,
  kappaA,
  kappaB,
  invGamma,
  datTheta,
  datMu,
  datProportion,
  datEvent,
  datTime
)
}
\arguments{
\item{n}{Number of samples to draw}

\item{nsamp}{How many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}

\item{dirichlet}{Not yet implemented}
}
\description{
Univariate ARMS for kappa
}
