% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_phi}
\alias{arms_phi}
\title{Univariate ARMS for phi}
\usage{
arms_phi(
  n,
  nsamp,
  ninit,
  minRange,
  maxRange,
  metropolis,
  simple,
  convex,
  npoint,
  currentPars,
  Delta,
  datProportion,
  datProportionConst
)
}
\arguments{
\item{n}{Number of samples to draw}

\item{nsamp}{How many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}

\item{dirichlet}{Not yet implemented}
}
\description{
Univariate ARMS for phi
}
