% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ars}
\alias{ars}
\title{Rcpp wrapper for ARS algorithm}
\usage{
ars(
  n,
  initialPoints,
  minRange,
  maxRange,
  jj,
  xis,
  vA,
  vB,
  datX0,
  datProportion,
  datEvent,
  weibullS
)
}
\arguments{
\item{n}{number of variates to draw}

\item{initialPoints}{at least 2 points in support to seed algorithm}

\item{minRange}{lower bound of support}

\item{maxRange}{upper bound of support}

\item{log_dens}{log density function, a function with a single parameter}
}
\description{
Rcpp wrapper for ARS algorithm
}
