% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ars_debug}
\alias{ars_debug}
\title{Duplicate wrapper for ARS algorithm for debugging. This version
returns the algorithm state as well as the variates.}
\usage{
ars_debug(
  n,
  initialPoints,
  minRange,
  maxRange,
  maxiter,
  jj,
  xis,
  vA,
  vB,
  datX0,
  datProportion,
  datEvent,
  weibullS
)
}
\arguments{
\item{n}{number of variates to draw}

\item{initialPoints}{at least 2 points in support to seed algorithm}

\item{minRange}{lower bound of support}

\item{maxRange}{upper bound of support}

\item{maxiter}{maximum number of iterations; zero if no limit}

\item{log_dens}{log density function, a function with a single parameter}
}
\description{
Duplicate wrapper for ARS algorithm for debugging. This version
returns the algorithm state as well as the variates.
}
