% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_support.R
\name{convex_support}
\alias{convex_support}
\title{Convex support}
\usage{
convex_support(x)
}
\arguments{
\item{x}{}
}
\value{
An object of ...
}
\description{
This is an indicator function of the convex support of the target density. This is required by R-pkg 'HI'
}
\examples{

x <- 1

}
