% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_support_kappas.R
\name{convex_support_kappas}
\alias{convex_support_kappas}
\title{Convex support}
\usage{
convex_support_kappas(x)
}
\arguments{
\item{x}{}
}
\value{
An object of ...
}
\description{
This is an indicator function of the convex support of Weibull's kappa
}
\examples{

x <- 1

}
