% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_support_phi.R
\name{convex_support_phi}
\alias{convex_support_phi}
\title{Convex support}
\usage{
convex_support_phi(x)
}
\arguments{
\item{x}{TBA}
}
\value{
An object of ...
}
\description{
This is an indicator function of the convex support of Dirichlet's dispersion phi
}
\examples{

x <- 1

}
