% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-likelihood}
\usage{
loglikelihood(par)
}
\arguments{
\item{par}{TBA}
}
\value{
An object of ...
}
\description{
Log-likelihood GTPCM
}
\examples{

x <- 1

}
