% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_beta_jl.R
\name{logpost_beta_jl}
\alias{logpost_beta_jl}
\title{Log posterior}
\usage{
logpost_beta_jl(x, ...)
}
\arguments{
\item{x}{}
}
\value{
An object of ...
}
\description{
This is a log posterior of each coefficients in the noncure fraction of GPTCM
}
\examples{

x <- 1

}
