% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_beta_l.R
\name{logpost_beta_l}
\alias{logpost_beta_l}
\title{Log posterior}
\usage{
logpost_beta_l(x)
}
\arguments{
\item{x}{proposal value}
}
\value{
An object of ...
}
\description{
This is a log posterior of all coefficients of one component in the noncure fraction of GPTCM. (This seems worse for estimating cure fraction's effects, maybe due to non-concavity)
}
\examples{

x <- 1

}
