% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_kappas.R
\name{logpost_kappas}
\alias{logpost_kappas}
\title{Log posterior}
\usage{
logpost_kappas(x)
}
\arguments{
\item{x}{proposal value}
}
\value{
An object of ...
}
\description{
This is a log posterior of Weibull's kappa
}
\examples{

x <- 1

}
