% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_phi.R
\name{logpost_phi}
\alias{logpost_phi}
\title{Log posterior}
\usage{
logpost_phi(x, ...)
}
\arguments{
\item{x}{}
}
\value{
An object of ...
}
\description{
This is a log posterior of the dispersion parameter phi of the Dirichlet distribution
}
\examples{

x <- 1

}
