% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_xi.R
\name{logpost_xi}
\alias{logpost_xi}
\title{Log posterior}
\usage{
logpost_xi(x)
}
\arguments{
\item{x}{proposal value}
}
\value{
An object of ...
}
\description{
This is a log posterior of each coefficients (including intercept) in the cure fraction of GPTCM. (This is slightly less efficient than 'logpost_xi2')
}
\examples{

x <- 1

}
