% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_xi2.R
\name{logpost_xi2}
\alias{logpost_xi2}
\title{Log posterior}
\usage{
logpost_xi2(x)
}
\arguments{
\item{x}{proposal value}
}
\value{
An object of ...
}
\description{
Log posterior of all coefficients (including intercept) in the cure fraction of GPTCM
}
\examples{

x <- 1

}
