% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_zeta_jl.R
\name{logpost_zeta_jl}
\alias{logpost_zeta_jl}
\title{Log posterior}
\usage{
logpost_zeta_jl(x, ...)
}
\arguments{
\item{x}{}
}
\value{
An object of ...
}
\description{
Log posterior of the proportion-related coefficients in the Dirichlet measure error model
}
\examples{

x <- 1

}
