% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_sampler.R
\name{metropolis_sampler}
\alias{metropolis_sampler}
\title{Metropolis sampler for a target density}
\usage{
metropolis_sampler(
  initial_value,
  n = n,
  proposal_shape = 1,
  proposal_scale = 1,
  theta = 1,
  proportion = 0.5,
  mu = 1,
  kappas = 0.9,
  burnin = 0,
  lag = 1
)
}
\arguments{
\item{initial_value}{}
}
\value{
An object of ...
}
\description{
Random number generator via MH algorithm. Adapted from https://blog.djnavarro.net/posts/2023-04-12_metropolis-hastings/
}
\examples{

x <- 1

}
