% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R, R/plotMCMC.R
\name{plotBrier}
\alias{plotBrier}
\alias{plotMCMC}
\title{Time-dependent Brier scores}
\usage{
plotBrier(
  dat,
  datMCMC,
  dat.new = NULL,
  time.star = NULL,
  xlab = "Time",
  ylab = "Brier score",
  ...
)

plotMCMC(dat, datMCMC, estimator = "xi")
}
\arguments{
\item{dat}{input data as a list containing survival data sub-list 
\code{survObj} with two vectors (\code{event} and \code{time}), clinical 
variable matrix \code{x0}, cluster-specific covariates \code{X}, and 
proportions data matrix \code{proportion}}

\item{datMCMC}{returned object from the main function \code{GPTCM()}}

\item{dat.new}{input data for out-sample prediction, with the same format 
as \code{dat}}

\item{time.star}{largest time for survival prediction}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{others}

\item{estimator}{print estimators, one of 
\code{c("beta", "zeta", "gamma", "eta")}}
}
\value{
A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.

A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.
}
\description{
Predict time-dependent Brier scores based on different survival models

MCMC
}
\examples{

x <- 1


x <- 1

}
