% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GPTCM.R
\name{predict.GPTCM}
\alias{predict.GPTCM}
\title{Predictions for a GPTCM}
\usage{
\method{predict}{GPTCM}(object, dat, newdata = NULL, type = "survival", ...)
}
\arguments{
\item{object}{the results of a \code{GPTCM} fit}

\item{dat}{the dataset used in \code{GPTCM()}}

\item{newdata}{optional new data at which to do predictions. If missing, the
prediction will be based on the training data}

\item{type}{the type of predicted value. Currently it is only valid with
'survival'}

\item{...}{for future methods}
}
\value{
A matrix object
}
\description{
Compute fitted values for a generalized promotion time cure model (GPTCM)
}
\examples{

x <- 1

}
