% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleTau.R
\name{sampleTau}
\alias{sampleTau}
\title{Gibbs sampling}
\usage{
sampleTau(n.sample, externalTauA, externalTauB, externalBetas)
}
\arguments{
\item{n.sample}{}
}
\value{
An object of ...
}
\description{
Gibbs sampling for the variance tauSq of nonzero betas
}
\examples{

x <- 1

}
