% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleV.R
\name{sampleV}
\alias{sampleV}
\title{Gibbs sampling}
\usage{
sampleV(n.sample, externalvA, externalvB, externalXi)
}
\arguments{
\item{n.sample}{}
}
\value{
An object of ...
}
\description{
Gibbs sampling for the variance v of xi
}
\examples{

x <- 1

}
