% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleW.R
\name{sampleW}
\alias{sampleW}
\title{Gibbs sampling}
\usage{
sampleW(n.sample, externalwA, externalwB, externalZetas)
}
\arguments{
\item{n.sample}{}
}
\value{
An object of ...
}
\description{
Gibbs sampling for the variance wSq of nonzero zetas
}
\examples{

x <- 1

}
