% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampling.R
\name{slice}
\alias{slice}
\title{Slice sampling}
\usage{
slice(
  n,
  init_theta,
  TARGET,
  w,
  max_steps,
  k_product,
  lower = -Inf,
  upper = +Inf,
  type = "raw",
  ...
)
}
\arguments{
\item{n}{specified number of samples to be generated}

\item{init_theta}{initial theta value}

\item{TARGET}{function proportional to what we want to sample from}

\item{w}{TBA}

\item{max_steps}{TBA}

\item{k_product}{TBA}

\item{lower}{TBA}

\item{upper}{TBA}

\item{type}{either "raw" or "log" dependet upon if TARGET is log(f(x))}

\item{...}{parameters passed to the TARGET function}
}
\value{
A vector of drawn values via slice sampler
}
\description{
This file will include functions which implement (product) slice sampling on
a target distribution adapted from
(i) https://github.com/TJHeaton/NonparametricCalibration/blob/main/SliceUpdateFunsFinal.R,
(ii) https://www.jarad.me/teaching/2013/10/29/slice-sampling-for-multimodal-TARGETs

We just need to call
slice(n, init_theta, TARGET, w, max_steps, k_product, lower, upper, type)
with TARGET being either:
1) Function f(x) which is proportional to the TARGET density
2) g(x) = log(f(x)) in which case need extra arg type = "log"
}
\examples{

x <- 1

}
