% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_gibbs_beta}
\alias{arms_gibbs_beta}
\title{Multivariate ARMS via Gibbs sampler for beta}
\usage{
arms_gibbs_beta(
  n,
  nsamp,
  ninit,
  minRange,
  maxRange,
  metropolis,
  simple,
  convex,
  npoint,
  currentPars,
  tauSq,
  kappa,
  datX,
  datTheta,
  datMu,
  datProportion,
  datEvent,
  datTime,
  weibullS
)
}
\arguments{
\item{n}{Number of samples to draw}

\item{nsamp}{How many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}
}
\description{
Multivariate ARMS via Gibbs sampler for beta
}
