% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arms_gibbs_xi}
\alias{arms_gibbs_xi}
\title{Multivariate ARMS via Gibbs sampler}
\usage{
arms_gibbs_xi(
  n,
  every,
  ninit,
  minRange,
  maxRange,
  metropolis,
  simple,
  convex,
  npoint,
  currentPars,
  vA,
  vB,
  datX0,
  datProportion,
  datEvent,
  weibullS
)
}
\arguments{
\item{n}{Number of samples to draw}

\item{every}{How many samples to draw for generating each sample; only the last 'n' draws will be kept; but it seems not work for every>1, not know why}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}
}
\description{
Multivariate ARMS via Gibbs sampler
}
