% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ars_gibbs}
\alias{ars_gibbs}
\title{Gibbs sampling for multivariate ARS}
\usage{
ars_gibbs(
  n,
  initialPoints,
  minRange,
  maxRange,
  xis,
  vA,
  vB,
  datX0,
  datProportion,
  datEvent,
  weibullS
)
}
\arguments{
\item{n}{number of variates to draw}

\item{initialPoints}{this can be a matrix in multivariate case}
}
\description{
Gibbs sampling for multivariate ARS
}
