% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPTCM.R
\name{GPTCM}
\alias{GPTCM}
\title{Fit Bayesian GPTCM Model}
\usage{
GPTCM(
  dat,
  n,
  p,
  L,
  proportion.model = TRUE,
  dirichlet = TRUE,
  method = "Bayes",
  hyperpar = NULL,
  kappaPrior = "IGamma",
  kappaSampler = "arms",
  w0Sampler = "IGamma",
  initial = NULL,
  nIter = 1,
  burnin = 0,
  thin = 1,
  tick = 100
)
}
\arguments{
\item{dat}{a list containing observed data from \code{n} subjects with
components \code{t}, \code{di}, \code{X}. For graphical learning of the
Markov random field prior, \code{survObj} should be a list of the list with
survival and covariates data. For subgroup models with or without graphical
learning, \code{survObj} should be a list of multiple lists with each
component list representing each subgroup's survival and covariates data}

\item{n}{TBA}

\item{p}{TBA}

\item{L}{TBA}

\item{proportion.model}{logical value; should the proportions be modeled or
not. If (\code{proportion.model = FALSE}), the argument \code{dirichlet} will
be invalid}

\item{dirichlet}{logical value; should the proportions be modeled via the
common (\code{dirichlet = FALSE}) or alternative (\code{dirichlet = TRUE})
parametrization of the Dirichlet regression model}

\item{method}{either `MLE` or `Bayes`}

\item{hyperpar}{TBA}

\item{kappaPrior}{TBA}

\item{kappaSampler}{TBA}

\item{w0Sampler}{TBA}

\item{initial}{TBA}

\item{nIter}{TBA}

\item{burnin}{TBA}

\item{thin}{TBA}

\item{tick}{TBA}
}
\value{
An object of ...
}
\description{
This is the main function to fit the Bayesian GPTCM model with multiscale data
for sparse identification of high-dimensional covariates.
}
\examples{

x <- 1

}
