% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R, R/plotMCMC.R
\name{plotBrier}
\alias{plotBrier}
\alias{plotMCMC}
\title{Time-dependent Brier scores}
\usage{
plotBrier(
  dat,
  datMCMC,
  time.star = NULL,
  xlab = "Time",
  ylab = "Brier score",
  ...
)

plotMCMC(dat, datMCMC, estimator = "xi")
}
\arguments{
\item{dat}{TBA}

\item{datMCMC}{TBA}

\item{...}{TBA}

\item{estimator}{TBA}
}
\value{
A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.

A \code{ggplot2::ggplot} object. See \code{?ggplot2::ggplot} for more
details of the object.
}
\description{
Predict time-dependent Brier scores based on different survival models

MCMC
}
\examples{

x <- 1


x <- 1

}
