% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeff.R
\name{plotCoeff}
\alias{plotCoeff}
\title{Plots for estimated coefficients}
\usage{
plotCoeff(
  dat,
  datMCMC,
  estimator = "beta",
  intercept = FALSE,
  bandwidth = NULL,
  xlim = NULL,
  xlab = NULL,
  label.y = NULL,
  first.coef = NULL,
  y.axis.size = 8,
  ...
)
}
\arguments{
\item{dat}{TBA}

\item{datMCMC}{TBA}

\item{estimator}{TBA}

\item{intercept}{TBA}

\item{bandwidth}{TBA}

\item{xlim}{TBA}

\item{xlab}{TBA}

\item{label.y}{TBA}

\item{first.coef}{TBA}

\item{y.axis.size}{TBA}

\item{...}{TBA}
}
\value{
An object of ...
}
\description{
create nice plots for estimated coefficients and 95% credible intervals
}
\examples{

x <- 1

}
