% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeff.R
\name{plotCoeff}
\alias{plotCoeff}
\title{Plots for estimated coefficients}
\usage{
plotCoeff(
  dat,
  datMCMC,
  estimator = "beta",
  label.y = NULL,
  legend.labs = NULL,
  ...
)
}
\arguments{
\item{dat}{TBA}

\item{datMCMC}{TBA}

\item{estimator}{TBA}

\item{...}{TBA}
}
\value{
An object of ...
}
\description{
create nice plots for estimated coefficients and 95% credible intervals
}
\examples{

x <- 1

}
