% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Target density}
\usage{
target(x, theta, proportion, mu, kappas)
}
\arguments{
\item{x}{TBA}

\item{theta}{TBA}

\item{proportion}{TBA}

\item{mu}{TBA}

\item{kappas}{TBA}
}
\value{
An object of ...
}
\description{
Predefined target density for S_pop in GTPCM
}
\examples{

x <- 1

}
